Development of V2
Added a new folder with the development of V2.

This is a fairly big update to fix a bug which resulted in the
calculation of DCI ignoring upstream connections.

If you use these versions please update me if there are any bugs ro
things are not working so I can fix them